feat(#200): Toast 디자인 수정 및 연관 불필요 코드 제거
