+++ src/app/layout.tsx
@@ -3,7 +3,7 @@ import "./globals.css";
 import { Analytics } from "@vercel/analytics/next";
 import { SpeedInsights } from "@vercel/speed-insights/next";
 
-import { ToastProvider } from "@/components/common/Toast";
+import ToastProvider from "@/components/common/Toast";
 import Overlay from "@/hooks/useOverlay/Overlay";
 import { pretendard } from "@/lib/fonts";
 import { initMocks } from "@/mocks";
@@ -25,10 +25,7 @@ export default function RootLayout({
       <body className={pretendard.className}>
         <MSWInitializer>
           <ReactQueryClientProvider>
-            <ToastProvider
-              position="bottom-right"
-              max={5}
-            >
+            <ToastProvider max={5}>
               <Overlay />
               {children}
             </ToastProvider>
